(NC-255) Refactor arguments in postgres storage.
Previously, the functions in classifier.storage.postgres were
inconsistent in taking either a direct db spec, or a Postgres record or
similarly shaped map with the spec under the :db key. Many of the ones
that took a Postgres record would destroy the record by destructuring it
and then constructing a new map to pass to other functions in the
namespace, erasing any protocol associations that the record had.

Fix this by changing all functions to take a Postgres record and pass
that record on to the other functions in the namespace that they call.