(NC-255) Storage -> {Primitive,Optimitized}Storage
Note for those who may be bisecting: this breaks the HTTP & Acceptance
tests! See the last paragraph below for more information.

Refactor the Storage protocol abstraction, which was starting to
encompass all of the app's functionality because of the plan to do RBAC
at the storage layer. Since some operations (e.g. classification)
require knowledge that the user may not be allowed to see, in order to
have the full information but still not leak things to the users, the
operation has to operate at the RBAC'd Storage layer, hence the
blob-like growth of the Storage protocol's functionality. Furthermore,
because the PermissionedStorage protocol wraps a Storage instance and
just filters the output, but needs to perform additional Storage
operations to get the information it needs in order to determine what to
filter, some operations that were moved into Storage resulted in many
additional database retrievals. Finally, some Storage methods could be
entirely implemented using other Storage methods, hence didn't need to
be repeated in each Storage implementation.

This commit breaks apart Storage into three pieces:
  1. PrimitiveStorage contains only those methods that can't be
     implemented using other Storage methods; it expresses the minimimum
     necessary functionality needed to be used for Classifier storage.

  2. OptimizedStorage contains methods that can be naively implemented
     using PrimitiveStorage methods, but may be implemented in a more
     efficient manner for particular storage implementations. Storage
     users are expected to check whether the instance they are using
     implements OptimizedStorage, and if not, use the naive
     implemantations. Since it is not possible to check whether an
     instance implements specific protocol methods, anything that
     implements OptimizedStorage must use the naive implementations for
     the OptimizedStorage methods that it does not itself implement.

  3. Some methods that were previously in the Storage protocol largely
     so that they could happen on the correct side of the RBAC wrapping
     in PermissionedStorage have been removed from the Storage layer
     entirely and have equivalents in the Classifier application
     protocol that was added in the previous commit.

Because the Storage protocol is now gone, and the API handler expects to
be given a storage instance, the HTTP tests and Acceptance tests are
totally borked right now, but all other tests should be passing.