(NC-255) Add & use Classifier ptcl implementation.
Add a default implementation of the Classifier protocol that was added
two commits ago.

Change the api handler to expect an instance of the Classifier, rather
than an instance of Storage. Accordingly, change the http tests and the
main namespace to supply one to the api handler (this fixes the http and
acceptance tests; all tests should be passing now).