(NC-212) Add get-group-ids to OptimizedStorage.
The RBAC API has changed since PermissionedStorage was written, and no
longer has the capability to list all group ids that a subject has
permission to perform a specific action on. Some of PermissionedStorage
was written with the assumption that that would be possible in order to
get a set of ids that the subject has permission to view, and now we
have to work around that.

The workaround will be to get *all* group ids, and then filter from
those the ones that the subject can view. Since fetching all fields of
all groups in order to just get the ids would be egregiously
inefficient, add a `get-group-ids` method to OptimizedStorage, and add
an implementation of it for the Postgres storage implementation.