(NC-212) Change signature of viewable-group-ids.
Change the signature of the `viewable-group-ids` permission function in
classifier.storage.permissioned to take a collection of all known group
ids, since the RBAC API that it uses no longer has the capability to
enumerate ids that the subject has permissions to perform a particular
action on; we must instead filter only the viewable ids from those
passed to the permission function.

Change storage-with-permissions to pass all known ids to the
viewable-group-ids permission function, and change the
PermissionedStorage tests' mock implementation of that fn to take
another argument for these ids.