(NC-259) Seperate config for client SSL cert.
The SSL certificate that the classifier uses when it connects to the
Puppet Master's REST API over SSL can now be configured seperately from
the certificate used for the API webserver. The configuration keys are
all the same, they just go in the top-level `classifier` configuration
object, rather than the one nested under `webserver`.