(NC-213) Add RBAC to classifier.
In main.clj, when the [:classifier :access-control] configuration key is
truthy, then create a PermissionedClassifier application instance to
pass to http/api-handler, and add authentication by wrapping the
resultant handler with RBAC's authentication middleware.

When the [:classifier :access-control] configuration key is not truthy,
then the classifier behaves as before (i.e. without any authentication
or authorization).

Add the RBAC services that the classifier uses for access control to the
classifier's bootstrap.cfg.

Add classifier and webserver HOCON config files to
./dev-resources/test-conf.d, and use them in the acceptance tests
instead of the configuration map literal.