(NC-213) Update integration tests for RBAC.
In the install_postgres method in the integration helper, setup
a database for the RBAC service to use.

In the 95_install_classifier_cert pre-suite "test", setup the RBAC
service's webserver (making it the default server), explicitly specify
the server id for the classifier service's handlers, and configure the
RBAC service's database.