(maint) Update pe-rbac-service dep to 0.1.12
In order to add an integration test that ensures that the RBAC
permission checking is working correctly, we need to update RBAC to
a version that has the API user (which is the subject when directly
accessing the API via cert whitelists). Also new in this version is the
switch to UUIDs for user/subject ids, which requires changing some
fixtures.

Change the date-time serialization format to include milliseconds in its
ISO8601 output, in order to be consistent with RBAC.