(NC-264) Avoid liberator representation clashes.
Liberator uses multimethod extension in order allow for custom
serializations to particular mime types. Due to the implicit global
state in multimethod extension, if multiple consumers of liberator
extend the multimethod to the same mime type, only one will win.

The puppetlabs/liberator-util library allows the consumers to coordinate
their overrides by extending the multimethod with an implementation that
tries to find an encoder in a dynamic variable. Each consumer can set
their encoder for the dynamic scope of their application (e.g. using
a ring middleware), avoiding conflict.

Remove the liberator defmethods from http.clj and add middleware that
sets liberator-util's map and seq encoders for `application/json` to
the classifier's custom json encoder, and add those middleware to the
standard middleware package.

In main.clj and the http-test, install liberator-util's defmethods for
liberator's representation multimethods that looks in the dynamic var
for the encoders that we set.