(maint) Less brittle group->classification.
Instead of removing a hardcoded list of keys from the passed group,
change group->classification to select from the group only the keys that
are present in the Classifier schema. Also, add schema annotations to
the function.

Due to adding schema-enforcement on group->classification, change
postgres/import-tree to pass ancestor groups instead of ancestor
classifications to postgres/validate-classes-and-parameters (which does
the group->classification conversion itself).