(maint) Use pe-beaker when on our Gem mirror
pe-beaker is always the latest released beaker + any changes needed to
allow working with an unreleased PE version. It is needed in order to
run the classifier tests against unreleased PE versions.

The changes in pe-beaker are intended to be confined to pe-specific
bits, so I don't anticipate this causing issues when running the
classifier tests standalone.