(NC-266) Set node environment from request when agent-specified
After retrieving an environment classification of 'agent-specified', the
node classifier indirector would remove the environment setting from the
retrieved result, causing the new node object to fall back to defaults
for its environment. Ultimately a node with a nil environment ends up
looking up the Puppet[:environment] which is unlikely to be the
environment from the original agent request.  However, because the
integration test agents were requesting the environment 'production',
the test was passing because this is the default Puppet[:environment] on
the master as well.

This commit now sets the node's environment to the request.environment
if the classifier indicates 'agent-specified', essentially falling back
to the plain node_terminus behavior. The integration test is updated to
check that a non default agent specified environment 'testing' is correctly
 retrieved.