(PDB-897) Create new message listener TK service
This commit refactors our current command code to allow message
listeners to be register via TK service on startup. The new message
listener service handles spawning threads to listen to the queue and the
plumbing of pulling a message off the queue, putting it back on in the
case of failure etc. The new message listener service consumes the MQ
config from the PuppetDBServer service. The functionality of the server
service is the same, but previously didn't expose a protocol. It now
needs to expose one so that the message listener service can be sure it
has started before it starts. The code handling the commands registers
itself with the new message listener service but is unchanged beyond
that.