(PDB-2625) Upgrade raynes.fs to 1.4.6
This harmonizes the version of raynes.fs with the one used in other
projects.

This requires upgrading kitchensink and trapperkeeper as well, because
raynes.fs 1.4.6 is backward-incompatible with 1.4.5. Specifically, the
absolute-path function is changed and the normalized-path function is
renamed. The old version of absolute-path has been added to kitchensink
1.3.0, so we now use that. Because trapperkeeper 1.2.0 also depended on
fs 1.4.5 (and is thus broken by upgrading it), it's also now version
1.3.1.